+++ src/main/java/at/hannibal2/skyhanni/data/RenderGuiData.kt
@@ -15,18 +15,7 @@ class RenderGuiData {
     fun onRenderOverlay(event: RenderGameOverlayEvent.Pre) {
         if (event.type != RenderGameOverlayEvent.ElementType.HOTBAR) return
 
-//        val currentScreen = Minecraft.getMinecraft().currentScreen
-//        if (currentScreen != null) {
-//            if (currentScreen is GuiInventory || currentScreen is GuiChest) return
-//        }
-
-        GlStateManager.pushMatrix()
-        GlStateManager.translate(0f, 0f, -1f)
-        GlStateManager.enableDepth()
-
         GuiRenderEvent.GameOverlayRenderEvent().postAndCatch()
-
-        GlStateManager.popMatrix()
     }
 
     @SubscribeEvent