+++ CHANGELOG.md
@@ -35,6 +35,9 @@
 + Added **Optimal Speed** - Show the optimal speed for your current tool in the hand. (Ty MelonKingDE for the values)
 + Added **Warn When Close** - Warn with title and sound when the next crop milestone upgrade happens in 5 seconds. Useful for switching to a different pet for leveling.
 + Added **Money per Hour** - Displays the money per hour YOU get with YOUR crop/minute value when selling the items to bazaar.
++ Added farming contest timer.
++ Added wrong fungi cutter mode warning.
++ Added show the price per garden experience inside the visitor gui.
 
 
 ### Features from other Mods